+++ Sources/PCBasicCursor.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+public struct PCBasicCursor {
+    public let type: PCCursorType
+    public let position: Int
+    public let roomId: Int
+    public let updatedAt: String
+    public let userId: String
+}